requirements.txt
api.video==1.4.7
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
boto3==1.38.27
botocore==1.38.27
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cloudinary==1.44.0
colorama==0.4.6
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.2.0
fastjsonschema==2.21.1
ffmpeg==1.4
ffmpeg-python==0.2.0
Flask==3.1.0
flask-cors==6.0.0
future==1.0.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
moviepy==2.2.1
mysql-connector-python==9.2.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.2.6
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==11.2.1
pipreqs==0.5.0
platformdirs==4.3.8
proglog==0.1.12
prompt_toolkit==3.0.51
pure_eval==0.2.3
Pygments==2.19.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pywin32==310
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
s3transfer==0.13.0
six==1.17.0
soupsieve==2.7
stack-data==0.6.3
tinycss2==1.4.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.14.0
urllib3==2.4.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
yarg==0.1.9
yt-dlp==2025.5.22